requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
itsdangerous==2.1.2
Click==8.1.7
Jinja2==3.1.2

pymodbus==3.6.3
pyserial==3.5

termux-api==0.5

# Optional utilities
requests==2.31.0
